pyproject.toml
[project]
name = "rtb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.4.1",
    "crewai>=0.193.2",
    "crewai-tools>=0.76.0",
    "faiss-cpu>=1.13.2",
    "huggingface-hub>=0.36.0",
    "langchain>=1.2.6",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.7",
    "langchain-google-genai>=4.2.0",
    "langchain-huggingface>=1.2.0",
    "langchain-text-splitters>=1.1.0",
    "pypdf>=6.6.0",
    "requests>=2.32.5",
    "sentence-transformers>=5.2.0",
    "tavily-python>=0.7.19",
]
